+++ src/products/components/ProductVariants/ProductVariants.tsx
@@ -355,8 +355,8 @@ export const ProductVariants: React.FC<ProductVariantsProps> = props => {
                     {isDefault && (
                       <span className={classes.defaultVariant}>
                         {intl.formatMessage({
-                          defaultMessage: "default",
-                          description: "default variant label"
+                          defaultMessage: "Default",
+                          description: "default product variant indicator"
                         })}
                       </span>
                     )}
@@ -399,7 +399,7 @@ export const ProductVariants: React.FC<ProductVariantsProps> = props => {
                     data-test="actions"
                     onClick={e => e.stopPropagation()}
                   >
-                    {variant?.id !== product?.defaultVariant.id && (
+                    {variant?.id !== product?.defaultVariant?.id && (
                       <ProductVariantSetDefault
                         onSetDefaultVariant={() => onSetDefaultVariant(variant)}
                       />